http://www.mjret.in/V3I1/M2-3-1-1-2016.pdf
Friends invitation interactions among users as a social graph. Based on this social graph we
proposed two key methods in order to detect fake identities that are fake users. Time
Required to Browse and the time Gap Between the posting the status, sending SMS etc.
Method is a voting-based Sybil detection and second is Sybil community detection to find
other colluding Sybil around identified Sybil. In the second method we are using the global
acceptance rate in order to detect the fake users. We are going to show the results in the
form of global acceptance ratio of a particular user.

Keywords: Online Social Network, Fake Identities, Global Acceptance ratio, Sybil attacks, social
networks, social network-based Sybil defense, communities.